Arrays, tuples, structs, local type inference, && and ||, static and main

+++ src/drlc/low/drc1/instruction/jump/InstructionConditionalJump.java
@@ -4,7 +4,7 @@ import drlc.low.drc1.instruction.Instruction;
 
 public abstract class InstructionConditionalJump extends InstructionJump {
 	
-	public InstructionConditionalJump(int section) {
+	public InstructionConditionalJump(short section) {
 		super(section);
 	}
